bug(Rotation): Fix shape rotation no longer syncing

+++ server/src/api/models/shape/position.py
@@ -4,7 +4,7 @@ from ..helpers import TypeIdModel
 
 
 class ShapePosition(BaseModel):
-    angle: int
+    angle: float
     points: list[tuple[float, float]]
